Feat: 딜 스코어 double 변경, 소숫점 버리고 Int형으로 response

+++ src/main/java/com/cherrypick/backend/domain/deal/service/RecommenderService.java
@@ -268,7 +268,7 @@ public class RecommenderService
                     DealService.getInfoTags(deal),
                     deal.getPrice(),
                     deal.getCreatedAt().toString(),
-                    deal.getDealScore(),
+                    (int) deal.getDealScore(),
                     (int) likeCount,
                     (int) commentCount,
                     deal.isSoldOut()
@@ -343,7 +343,7 @@ public class RecommenderService
                     DealService.getInfoTags(deal),
                     deal.getPrice(),
                     deal.getCreatedAt().toString(),
-                    deal.getDealScore(),
+                    (int) deal.getDealScore(),
                     (int) likeCount,
                     (int) commentCount,
                     deal.isSoldOut(),
